refactor: Rust-idiomatic API cleanup (methods, typed returns, encapsulation)
- DRED: encapsulate fields (pub→pub(crate)), add create/parse/process
  methods on OpusDREDDecoder, remove free function exports from lib.rs
- Packet utilities: typed returns (Bandwidth, Channels, Result<usize>)
  with internal _raw variants for codec internals
- Rename get_size→size on MappingMatrix, OpusMSEncoder, OpusMSDecoder,
  OpusProjectionEncoder, OpusProjectionDecoder; surround_get_size→surround_size
- Add set_signalling() on OpusCustomEncoder/OpusCustomDecoder
- Add is_loaded() accessor on OpusDREDDecoder

+++ tests/opus_decode.rs
@@ -181,7 +181,7 @@ fn test_decoder_all_2byte_prefixes() {
         packet[2] = 255;
 
         // Verify channel count from packet header
-        let nb_channels = opus_packet_get_nb_channels(packet[0]);
+        let nb_channels = i32::from(opus_packet_get_nb_channels(packet[0]));
         assert_eq!(
             nb_channels,
             (i & 1) + 1,
